feat: create tollkitdisplay component
